start on updating mupdf

+++ mupdf/source/fitz/crypt-arc4.c
@@ -24,10 +24,10 @@
 #include "mupdf/fitz.h"
 
 void
-fz_arc4_init(fz_arc4 *arc4, const unsigned char *key, unsigned keylen)
+fz_arc4_init(fz_arc4 *arc4, const unsigned char *key, size_t keylen)
 {
 	unsigned int t, u;
-	unsigned int keyindex;
+	size_t keyindex;
 	unsigned int stateindex;
 	unsigned char *state;
 	unsigned int counter;
@@ -86,9 +86,9 @@ fz_arc4_next(fz_arc4 *arc4)
 }
 
 void
-fz_arc4_encrypt(fz_arc4 *arc4, unsigned char *dest, const unsigned char *src, unsigned len)
+fz_arc4_encrypt(fz_arc4 *arc4, unsigned char *dest, const unsigned char *src, size_t len)
 {
-	unsigned int i;
+	size_t i;
 	for (i = 0; i < len; i++)
 	{
 		unsigned char x;
